requirements.txt
# ─────────────────────────────────────────────────────────────
#  Fury AI Voice Assistant  |  requirements.txt
#  © 2026 Fayaz Ahmed Shaik. All rights reserved.
# ─────────────────────────────────────────────────────────────

# ── Web Framework ─────────────────────────────────────────────
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# ── Text-to-Speech ────────────────────────────────────────────
edge-tts

# ── AI / LLM + Whisper STT (Groq — completely free tier) ─────
# Groq provides both LLM (Llama 3) and Whisper STT via cloud API.
# No PyTorch or local model downloads needed — keeps memory tiny.
groq

# ── Environment variable management ──────────────────────────
python-dotenv
